Remove graph display, remove empty Config pgae, remove dependencies on d3 and grapher.
