+++ misc/web/index.html
@@ -32,56 +32,11 @@
 					<i class="fa fa-server fa-2x"></i>
 					<span>Hosts</span>
 				</a>
-				<a class="nav-yellow" href="#config">
-					<i class="fa fa-cogs fa-2x"></i>
-					<span>Config</span>
-				</a>
 			</nav>
 		</div>
 
 		<div id="content">
 			<div id="vms">
-				<div id="vms-network" class="box">
-					<h1 class="box-header">Network Graph</h1>
-					<div class="box-content">
-						<div id="chart">
-							<div id="chart-toolbar" class="btn-group" role="group">
-								<button type="button" class="btn btn-default" id="center-button">
-									<i class="fa fa-compress"></i>
-									<span>Center</span>
-								</button>
-								<button type="button" class="btn btn-default" id="reflow-button">
-									<i class="fa fa-refresh"></i>
-									<span>Reflow</span>
-								</button>
-							</div>
-						</div>
-						<div id="sidebar-container">
-							<div id="sidebar">
-								<div id="node-data" class="uninitialized">
-									<div id="node-heading">
-										<h1 id="node-name"></h1>
-										<span id="node-count"></span>
-									</div>
-									<ul id="subnodes">
-									</ul>
-									<div id="node-info">
-										<table class="graph-table">
-										</table>
-									</div>
-								</div>
-							</div>
-							<div id="popup" class="sidepoint hidden">
-								<h1 id="machine-name">vm-foo</h1>
-								<div id="machine-info">
-									<table class="graph-table">
-									</table>
-								</div>
-							</div>
-						</div>
-					</div>
-				</div>
-
 				<div id="vms-table" class="box">
 					<h1 class="box-header">VM List</h1>
 					<div class="box-content">
@@ -116,29 +71,13 @@
 				</div>
 			</div>
 
-			<div id="config">
-				<div id="config-box" class="box">
-					<h1 class="box-header">Configuration</h1>
-					<div class="box-content">
-						<div style="text-align:center;font-style:italic;opacity:0.5;font-size:1.5em">Nothing here... yet.</div>
-					</div>
-				</div>
-			</div>
 		</div>
 
 		<script type="text/javascript" src="./libs/jquery-1.11.1.min.js"></script>
-		<script type="text/javascript" src="./libs/d3.min.js"></script>
-
-		<script type="text/javascript" src="./libs/grapher/grapher.js"></script>
-		<script type="text/javascript" src="./libs/grapher/center.js"></script>
-		<script type="text/javascript" src="./libs/grapher/palette.js"></script>
-		<script type="text/javascript" src="./libs/grapher/target.js"></script>
-		<script type="text/javascript" src="./libs/grapher/zoom.js"></script>
 
 		<script type="text/javascript" src="./libs/dataTables/jquery.dataTables.js"></script>
 		<script type="text/javascript" src="./libs/dataTables/dataTables.bootstrap.js"></script>
 
-		<script type="text/javascript" src="./js/graph.js"></script>
 		<script type="text/javascript" src="./js/glue.js"></script>
 	</body>
 </html>
